[bees] Surface / Detail tab layout in hivetool
## What
Extract the hivetool ticket detail view into a tabbed `Surface · Detail` layout, so the agent's curated surface gets a dedicated full-pane view.

## Why
Phase 5a of Project Surface. The surface was previously embedded as a block inside the timeline. Giving it its own tab makes it the primary presentation and leaves the timeline as a secondary diagnostic view.

## Changes

### `hivetool/src/ui/`
- **[NEW] `ticket-pane.ts`** — wrapper component owning the ticket header (title, status badge, identity chips, task controls) and a `Surface · Detail` tab bar. Probes for `surface.json` on ticket change; hides tabs when no surface exists.
- **[NEW] `surface-pane.ts`** — full-pane "Surface" tab body. Owns surface loading and renders `<bees-surface-view>`.
- **[MODIFIED] `ticket-detail.ts`** — stripped header, identity chips, surface loading, and task controls. Now a pure timeline renderer.
- **[MODIFIED] `app.ts`** — swapped `<bees-ticket-detail>` for `<bees-ticket-pane>` in the tickets tab main area.

### `PROJECT_SURFACE.md`
- Phase 5a items checked off.

## Testing
1. Run `npm run dev:hivetool -w packages/bees`.
2. Select a ticket **without** a `surface.json` — no tab bar, detail view renders directly.
3. Select a ticket **with** a `surface.json` (e.g. `37eea4e3`) — `Surface · Detail` tabs appear, Surface is auto-selected.
4. Switch between tabs — each gets the full main pane.
